[refactor] 리뷰반영
DoubleBackPressToExit 분리, Scaffold bottomBar 사용
