+++ core/common/src/main/res/values/strings.xml
@@ -27,4 +27,8 @@
     <string name="sign_in_dialog">로그인이 필요합니다</string>
     <string name="sign_in_dialog_dismiss">취소</string>
     <string name="google_id_no_credential_exception_message">기기에 로그인된 구글 계정이 없습니다</string>
+
+    <!-- Back to exit -->
+    <string name="back_to_exit">한 번 더 누르면 종료됩니다.</string>
+
 </resources>
