+++ feature/permission/src/main/res/values/strings.xml
@@ -14,14 +14,12 @@
         실제 오디오 녹음은 발생하지 않으며, 어떠한 음성 데이터도 저장하거나 전송하지 않습니다.
     </string>
 
-    <string name="back_to_exit">한 번 더 누르면 종료됩니다.</string>
-
     <!--Permission Bar-->
     <string name="permission_snack_bar_text">필수 권한을 허용해주세요.</string>
     <string name="action_to_setting">설정으로 이동</string>
 
     <string name="permission_location">위치</string>
     <string name="permission_location_content_desc">위치 권한</string>
-    <string name="permission_location_desc">지도의 내 위치에 음악을 등록하기 위해 위치권한이 필요합니다.</string>
+    <string name="permission_location_desc">지도의 현재 위치에 음악을 등록하기 위해 위치권한이 필요합니다.</string>
 
 </resources>
